requirements.txt
bcrypt==4.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
Django==4.0.3
django-cors-headers==3.11.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.0
packaging==22.0
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2022.6
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.13